+++ apps/web/src/App.tsx
@@ -6,6 +6,7 @@ import { CalendarView } from "./components/CalendarView";
 import { HelpOverlay } from "./components/HelpOverlay";
 import { LabelsManagement } from "./components/LabelsManagement";
 import { QuickAdd, type QuickAddHandle } from "./components/QuickAdd";
+import { AssistPanel } from "./components/AssistPanel";
 import { ReminderPoller } from "./components/ReminderPoller";
 import { SearchBox, type SearchBoxHandle } from "./components/SearchBox";
 import { Sidebar } from "./components/Sidebar";
@@ -185,6 +186,7 @@ function AppInner() {
             {navOpen ? "✕" : "☰"}
           </button>
           <QuickAdd ref={quickAddRef} view={view} projects={projects} />
+          <AssistPanel />
           <SearchBox ref={searchRef} onSelectTask={setSelectedTaskId} />
           <button
             type="button"